requirements.txt
@azure/search-documents
@azure/openai
@azure/identity
@azure/ai-documentintelligence
@azure/ai-contentsafety
@azure/storage-blob
@azure/cosmos
azure-monitor-query==1.2.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.1
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
structlog>=23.2.0
opencensus-ext-azure>=1.1.13
